Conflicts:
	pom.xml
	start-splice-dev.sh
	structured_constants/src/main/java/com/splicemachine/constants/SpliceConstants.java
	structured_constants/src/main/java/com/splicemachine/hbase/BufferedRegionScanner.java
	structured_constants/src/main/java/com/splicemachine/hbase/HBaseRegionLoads.java
	structured_derby/src/main/java/com/splicemachine/SpliceKryoRegistry.java
	structured_derby/src/main/java/com/splicemachine/derby/hbase/SpliceOperationRegionScanner.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/job/index/PopulateIndexTask.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/job/scheduler/JobControl.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/operations/BroadcastJoinOperation.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/operations/OperationTree.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/operations/SortOperation.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/operations/TableScanOperation.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/storage/DistributedClientScanProvider.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/storage/RegionAwareScanner.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/store/access/base/SpliceController.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/store/access/base/SpliceScan.java
	structured_derby/src/main/java/com/splicemachine/derby/impl/store/access/btree/IndexController.java
	structured_derby/src/main/java/com/splicemachine/derby/utils/SpliceAdmin.java
	structured_derby/src/test/java/com/splicemachine/derby/impl/job/load/HdfsImportIT.java
	structured_derby/src/test/java/com/splicemachine/derby/impl/sql/execute/operations/FixedSITableScannerTest.java
	structured_derby/src/test/java/com/splicemachine/derby/impl/sql/execute/tester/DataTypeCorrectnessIT.java
	structured_derby/src/test/java/org/apache/derby/impl/sql/execute/operations/InsertOperationIT.java
	structured_si/src/main/java/com/splicemachine/si/data/api/SDataLib.java
	structured_si/src/main/java/com/splicemachine/si/data/hbase/HDataLib.java
	structured_si/src/main/java/com/splicemachine/si/data/light/LDataLib.java
	structured_si/src/main/java/com/splicemachine/si/impl/FilterState.java

+++ structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/operations/MergeSortJoinRows.java
@@ -99,12 +99,12 @@ public class MergeSortJoinRows implements IJoinRowsIterator<ExecRow> {
     }
 
     @Override
-    public int getLeftRowsSeen() {
+    public long getLeftRowsSeen() {
         return leftRowsSeen;
     }
 
     @Override
-    public int getRightRowsSeen() {
+    public long getRightRowsSeen() {
         return rightRowsSeen;
     }
 }
